+++ code/src/cljs/sixsq/nuvla/ui/notifications/views.cljs
@@ -13,6 +13,7 @@
     [sixsq.nuvla.ui.notifications.events :as events]
     [sixsq.nuvla.ui.notifications.spec :as spec]
     [sixsq.nuvla.ui.notifications.subs :as subs]
+    [sixsq.nuvla.ui.notifications.utils :as utils]
     [sixsq.nuvla.ui.panel :as panel]
     [sixsq.nuvla.ui.utils.form-fields :as ff]
     [sixsq.nuvla.ui.utils.general :as general-utils]
@@ -266,10 +267,10 @@
     (fn []
       (let [header (str/capitalize (str (@tr [:edit]) " " (@tr [:subscription])))
             {:keys [name description method-ids enabled category resource-id resource-kind]} @subscription]
-        [:div]
         [ui/Modal {:open       @visible?
                    :close-icon true
                    :on-close   #(dispatch [::events/close-edit-subscription-modal])}
+
 
          [uix/ModalHeader {:header header}]
 
@@ -325,12 +326,15 @@
 
 
 (def criteria-metric-options
-  {"nuvlabox"               [{:key "cpu load" :text "CPU load %" :value "load"}
-                             {:key "ram usage" :text "RAM usage %" :value "ram"}
-                             {:key "disk usage" :text "Disk usage %" :value "disk"}
-                             {:key "state" :text "NuvlaEdge online" :value "state"}]
-   "infrastructure-service" [{:key "status" :text "status" :value "status"}]
-   "data-record"            [{:key "content-type" :text "content-type" :value "content-type"}]})
+  {"nuvlabox"               [{:key "cpu load" :text "CPU load %" :value utils/cpu-load }
+                             {:key "ram usage" :text "RAM usage %" :value utils/ram }
+                             {:key "disk usage" :text "Disk usage %" :value utils/disk }
+                             {:key "state" :text "NuvlaEdge online" :value utils/state }
+                             {:key "network rx" :text "Network Rx GB" :value utils/network-rx }
+                             {:key "network tx" :text "Network Tx GB" :value utils/network-tx }
+                             ]
+   "infrastructure-service" [{:key "status" :text "status" :value utils/status}]
+   "data-record"            [{:key "content-type" :text "content-type" :value utils/content-type}]})
 
 
 (def criteria-conditions
@@ -344,6 +348,8 @@
   {"nuvlabox"               {:load  :numeric
                              :ram   :numeric
                              :disk  :numeric
+                             :network-rx :numeric
+                             :network-tx :numeric
                              :state :boolean}
    "infrastructure-service" {:status :set}
    "data-record"            {:content-type :string}})
@@ -353,6 +359,8 @@
   {"nuvlabox"               {:load  (map (fn [x] {:key x :value x :text x}) [">" "<"])
                              :ram   (map (fn [x] {:key x :value x :text x}) [">" "<"])
                              :disk  (map (fn [x] {:key x :value x :text x}) [">" "<"])
+                             :network-rx (map (fn [x] {:key x :value x :text x}) [">"])
+                             :network-tx (map (fn [x] {:key x :value x :text x}) [">"])
                              :state ((get-in criteria-condition-type ["nuvlabox" :state])
                                      criteria-conditions)}
    "infrastructure-service" {:status ((get-in criteria-condition-type ["infrastructure-service" :status])
@@ -363,7 +371,8 @@
 
 (def component-options
   [{:key "nuvlabox", :text "NuvlaEdge Telemetry", :value "nuvlabox"}
-   {:key "infrastructure-service", :text "Infrastructure Service", :value "infrastructure-service"}
+   ;; FIXME: Enable creation of IS notification configurations when IS starts producing metrics.
+   #_{:key "infrastructure-service", :text "Infrastructure Service", :value "infrastructure-service"}
    {:key "data-record", :text "Data Record", :value "data-record"}])
 
 
@@ -401,6 +410,9 @@
                                       (fn [value] (swap! value-options #(conj @value-options value))))
                     :options        options}])))
 
+(def metric-name->default-condition
+  {utils/network-rx ">"
+   utils/network-tx ">"})
 
 (defn ConditionRow
   [metric-name condition collection on-change validate-form?]
@@ -409,12 +421,149 @@
                   :style      {:padding-bottom 8}} "Condition"]
    [ui/TableCell
     [ui/Dropdown {:selection true
+                  :disabled (boolean (metric-name->default-condition metric-name))
                   :options   (vec
                                ((keyword metric-name)
                                 (get criteria-condition-options collection)))
                   :error     (and validate-form? (not (seq condition)))
+                  :value     (or (metric-name->default-condition metric-name) condition)
                   :on-change (ui-callback/value
                                #(on-change :criteria {:condition %}))}]]])
+
+(def metric-name->use-other-translation-key
+  {utils/disk [:subs-notif-disk-use-other]
+   utils/network-rx [:subs-notif-network-use-other]
+   utils/network-tx [:subs-notif-network-use-other]})
+
+(def metrics-with-customizable-dev-name #{utils/network-rx utils/network-tx utils/disk})
+
+(defn- DeviceNameOptions
+  [{disabled? :disabled?}]
+  (let [tr (subscribe [::i18n-subs/tr])
+        criteria (subscribe [::subs/criteria])
+        use-other-than-default? (r/atom (or (:dev-name @criteria) false))]
+    (fn []
+      (let [metric-name (:metric @criteria)]
+        (when (metrics-with-customizable-dev-name metric-name)
+          [ui/TableCell {:col-span 2
+                         :class "font-weight-400"}
+           [:div {:style {:display :flex
+                          :justify-items :stretch-between
+                          :align-items :center
+                          :height 40
+                          :gap 24}}
+            [ui/Checkbox {:style {:margin-right 2}
+                          :label (some->> (metric-name->use-other-translation-key metric-name) (@tr))
+                          :default-checked @use-other-than-default?
+                          :read-only disabled?
+                          :on-change (ui-callback/checked
+                                      (fn [checked?]
+                                        (when (not checked?)
+                                          (dispatch [::events/remove-custom-name]))
+                                        (reset! use-other-than-default? checked?)))}]
+            (when @use-other-than-default?
+              [ui/Input {:type :text
+                         :placeholder (@tr [(keyword (str "subs-notif-name-of-" metric-name "-to-monitor"))])
+                         :name :other-disk-name
+                         :read-only disabled?
+                         :default-value (or (:dev-name @criteria) "")
+                         :on-change (ui-callback/value #(dispatch [::events/update-custom-device-name %]))
+                         :style {:flex 1}}])]])))))
+
+
+(defn- ResetIntervalOptions
+  [{disabled? :disabled?}]
+  (let [tr                   (subscribe [::i18n-subs/tr])
+        criteria             (subscribe [::subs/criteria])
+        validate-form?       (subscribe [::subs/validate-form?])]
+    (fn []
+      (let [reset-interval (:reset-interval @criteria)
+            monthly-reset? (or (nil? reset-interval) (= reset-interval "month"))
+            custom-reset? (not monthly-reset?)
+            start-date-of-month (or (:reset-start-date @criteria) 1)
+            custom-interval-days @(subscribe [::subs/custom-days])]
+        (when (utils/metrics-with-reset-windows (:metric @criteria))
+          [ui/TableCell {:col-span 2
+                         :class "font-weight-400"}
+           [:div {:style {:min-height 40}
+                  :class "grid-2-cols-responsive"}
+            [:div {:on-click #(dispatch (when  (not disabled?) [::events/choose-monthly-reset]))
+                   :style {:display :flex :align-items :center :opacity (if monthly-reset? 1 0.4)}}
+             [:input {:type :radio
+                      :name :reset
+                      :checked monthly-reset?
+                      :read-only true
+                      :id :monthly}]
+             [:label {:for :monthly
+                      :style {:margin-left "0.5rem" }} (@tr [:subs-notif-reset-on-day])]
+             [ui/Input {:type :number
+                        :error (and monthly-reset?
+                                    @validate-form?
+                                    (not (s/valid? ::spec/reset-start-date start-date-of-month)))
+                        :default-value start-date-of-month
+                        :disabled custom-reset?
+                        :read-only disabled?
+                        :style {:justify-self :start
+                                :margin-left "0.5rem"
+                                :max-width "100px"
+                                :margin-right "90px"}
+                        :label (@tr [:of-month])
+                        :label-position :right
+                        :on-change (ui-callback/value #(dispatch [::events/update-notification-subscription-config
+                                                                  :criteria
+                                                                  {:reset-start-date (js/Number %)}]))}]
+             [:div (ff/help-popup "min: 1, max: 31")]]
+            [:div {:on-click #(dispatch (when (not disabled?) [::events/choose-custom-reset]))
+                   :style {:display :flex :align-items :center :align-self "end" :opacity (if custom-reset? 1 0.4)}}
+             [:input {:type :radio
+                      :name :reset
+                      :checked custom-reset?
+                      :read-only true
+                      :id :custom}]
+             [:label {:for :custom
+                      :style {:margin-left "0.5rem"}} [:span (str/capitalize (@tr [:subs-notif-custom-reset-after]))]]
+             [ui/Input {:type :number
+                        :error (and custom-reset? @validate-form? (not (s/valid? ::spec/reset-interval reset-interval)))
+                        :default-value (or custom-interval-days 1)
+                        :disabled monthly-reset?
+                        :read-only disabled?
+                        :style {:justify-self :start
+                                :margin-left "0.5rem"
+                                :max-width "100px"
+                                :margin-right "60px"}
+                        :label (@tr [:days])
+                        :label-position :right
+                        :on-change (ui-callback/value #(dispatch [::events/update-custom-days %]))}]
+             [:div (ff/help-popup "min: 1, max: 999")]]]])))))
+
+(def default-start-date 1)
+(def default-custom-interval 1)
+
+(defn- get-network-info-text [criteria tr]
+  (let [start-day-of-month    (or (:reset-start-date criteria) default-start-date)
+        reset-in-days         (or (:reset-in-days criteria) default-custom-interval)
+        reset-interval-string (:reset-interval criteria)
+        interface-text        (if
+                               (str/blank? (:dev-name criteria))
+                                (@tr [:subs-notif-network-info-default])
+                                (@tr [:subs-notif-network-info-specific]))
+        interval-text         (if (or (str/blank? reset-interval-string) (= "month" reset-interval-string))
+                                (str (@tr [:subs-notif-network-reset-monthly]) " ("
+                                     (@tr [:subs-notif-reset-on-day]) " "
+                                     start-day-of-month ")")
+                                (str (@tr [:subs-notif-network-reset-custom]) " ("
+                                     (@tr [:subs-notif-custom-reset-after]) " "
+                                     reset-in-days " "
+                                     (@tr (if (= reset-in-days 1) [:day] [:days])) ")"))]
+    (str interface-text " " interval-text)))
+
+(defn- get-info-text [criteria tr]
+  (let [metric-name (:metric criteria)]
+    (case metric-name
+      utils/disk (when (str/blank? (:dev-name criteria)) (@tr [:subs-notif-disk-info]))
+      utils/network-tx (get-network-info-text criteria tr)
+      utils/network-rx (get-network-info-text criteria tr)
+      "")))
 
 
 (defn AddSubscriptionConfigModal
@@ -424,8 +573,7 @@
         validate-form?      (subscribe [::subs/validate-form?])
         form-valid?         (subscribe [::subs/form-valid?])
         on-change           (fn [name-kw value]
-                              (dispatch [::events/update-notification-subscription-config name-kw value])
-                              (dispatch [::events/validate-notification-subscription-config-form]))
+                              (dispatch [::events/update-notification-subscription-config name-kw value]))
         notif-methods       (subscribe [::subs/notification-methods])
         collection          (subscribe [::subs/collection])
         criteria-metric     (subscribe [::subs/criteria-metric])
@@ -433,7 +581,7 @@
         metric-name         (r/atom "")
         component-option    (r/atom "")
         value-options       (r/atom "")
-        subscription-config (subscribe [::subs/notification-subscription-config])]
+        subscription-config  (subscribe [::subs/notification-subscription-config])]
 
     (dispatch [::events/get-notification-methods])
     (dispatch [::events/reset-subscription-config-all])
@@ -510,15 +658,27 @@
              [ui/TableCell {:collapsing true
                             :style      {:padding-bottom 8}} "Metric"]
              [ui/TableCell
-              [ui/Dropdown {:selection true
-                            :options   (get criteria-metric-options @collection)
-                            :error     (and @validate-form? (not (seq @metric-name)))
-                            :on-change (ui-callback/value
+              [:div {:style {:display :flex
+                             :align-items :center
+                             :gap "0.3rem"}}
+               [ui/Dropdown {:selection true
+                             :options   (get criteria-metric-options @collection)
+                             :error     (and @validate-form? (not (seq @metric-name)))
+                             :on-change (ui-callback/value
                                          #(do
                                             (reset! metric-name %)
                                             (on-change
-                                              :criteria {:metric %
-                                                         :kind   (criteria-metric-kind @collection %)})))}]]]
+                                              :criteria (if (utils/metrics-with-reset-windows %)
+                                                          {:metric %
+                                                           :kind   (criteria-metric-kind @collection %)
+                                                           :reset-interval "month"
+                                                           :reset-start-date 1
+                                                           :condition ">"}
+                                                          {:metric %
+                                                           :kind   (criteria-metric-kind @collection %)}))))}]
+               [:div {:style  {:white-space :normal
+                               :font-size "0.9rem"}}
+                (get-info-text criteria tr)]]]]
 
             (case criteria-metric
               :string [:<>
@@ -561,7 +721,14 @@
                             :read-only   false
                             :on-change   (ui-callback/value #(on-change :criteria {:value %}))}]]]]
               :boolean nil
-              nil)]]
+              nil)
+
+            [ui/TableRow
+             [ResetIntervalOptions]]
+
+            [ui/TableRow
+             [DeviceNameOptions]]]
+           ]
 
           [ui/Header {:as "h3"} "Notification"]
           [ui/Form
@@ -584,8 +751,7 @@
         validate-form?      (subscribe [::subs/validate-form?])
         form-valid?         (subscribe [::subs/form-valid?])
         on-change           (fn [name-kw value]
-                              (dispatch [::events/update-notification-subscription-config name-kw value])
-                              (dispatch [::events/validate-notification-subscription-config-form]))
+                              (dispatch [::events/update-notification-subscription-config name-kw value]))
         notif-methods       (subscribe [::subs/notification-methods])
         criteria-metric     (subscribe [::subs/criteria-metric])
         components-number   (subscribe [::subs/components-number])
@@ -661,14 +827,20 @@
              [ui/TableCell {:collapsing true
                             :style      {:padding-bottom 8}} "Metric"]
              [ui/TableCell
-              [ui/Dropdown {:selection true
-                            :disabled  true
-                            :value     (:metric criteria)
-                            :options   (get criteria-metric-options collection)
-                            :on-change (ui-callback/value
+              [:div {:style {:display :flex
+                             :align-items :center
+                             :gap "0.3rem"}}
+               [ui/Dropdown {:selection true
+                             :disabled  true
+                             :value     (:metric criteria)
+                             :options   (get criteria-metric-options collection)
+                             :on-change (ui-callback/value
                                          #(on-change
                                             :criteria {:metric %
-                                                       :kind   (criteria-metric-kind collection %)}))}]]]
+                                                       :kind   (criteria-metric-kind collection %)}))}]
+               [:div {:style  {:white-space :normal
+                               :font-size "0.9rem"}}
+                (get-info-text criteria tr)]]]]
 
             (if-not (and (= collection "nuvlabox") (= (:metric criteria) "state"))
               [ui/TableRow
@@ -695,7 +867,11 @@
                   :name      "Value"
                   :read-only false
                   :value     (:value criteria)
-                  :on-change (ui-callback/value #(on-change :criteria {:value %}))}]]])]]
+                  :on-change (ui-callback/value #(on-change :criteria {:value %}))}]]])
+              [ui/TableRow
+               [ResetIntervalOptions {:disabled? true}]]
+              [ui/TableRow
+               [DeviceNameOptions {:disabled? true}]]]]
           [ui/Header {:as "h3"} "Notification"]
           [ui/Form
            [ui/FormGroup
@@ -873,17 +1049,18 @@
 (defn subscription-configs-table-header
   [tr]
   [ui/TableHeader
+   {:style {:font-weight 600}}
    [ui/TableRow
-    [ui/TableCell {:content ""}]
+    [ui/TableCell      [:span (str/capitalize (@tr [:name]))]
+     [:span ff/nbsp (ff/help-popup (@tr [:subscription-name]))]]
+    [ui/TableCell      [:span (str/capitalize (@tr [:criteria]))]
+     [:span ff/nbsp (ff/help-popup (@tr [:criteria-for-notifications]))]]
     [ui/TableCell
      [:span (str/capitalize (@tr [:enable]))]
      [:span ff/nbsp (ff/help-popup (@tr [:notifications-enable-disable-help]))]]
     [ui/TableCell
      [:span (str/capitalize (@tr [:notification-methods]))]
      [:span ff/nbsp (ff/help-popup (@tr [:notifications-methods-help]))]]
-    [ui/TableCell
-     [:span (str/capitalize (@tr [:subscriptions]))]
-     [:span ff/nbsp (ff/help-popup (@tr [:subscriptions-manage-help]))]]
     [ui/TableCell
      [:span (str/capitalize "action")]]]])
 
@@ -892,7 +1069,7 @@
   [subs-conf]
   (let [{:keys [metric condition kind value]} (:criteria subs-conf)]
     (r/as-element
-      [:span "criteria: " metric " "
+      [:span (str metric " ")
        [:span {:style {:font-weight "bold"}} condition]
        (when-not (= "boolean" kind) (str " " value))])))
 
@@ -913,8 +1090,7 @@
         subs-by-parent        (subscribe [::subs/subscriptions-by-parent])
         subs-by-parent-counts (subscribe [::subs/subscriptions-by-parent-counts])
         on-change             (fn [name-kw value]
-                                (dispatch-sync [::events/update-notification-subscription-config name-kw value])
-                                (dispatch [::events/validate-notification-subscription-config-form]))]
+                                (dispatch [::events/update-notification-subscription-config name-kw value]))]
     (dispatch [::events/get-notification-subscription-configs])
     (dispatch-sync [::events/get-notification-subscriptions])
     (dispatch [::events/get-notification-methods])
@@ -931,65 +1107,50 @@
                       [uix/Accordion
                        [ui/Table {:basic   "very"
                                   :compact true
+                                  :striped true
                                   :style   {:margin-top 10}}
 
                         [subscription-configs-table-header tr]
 
                         [ui/TableBody
-                         (doall (for [subs-conf resource-subs-confs]
-                                  ^{:key subs-conf}
-                                  [ui/TableRow
-                                   [ui/TableCell {:floated :left
-                                                  :width   2}
-                                    [:span (:name subs-conf)]
-                                    [:span ff/nbsp (ff/help-popup (criteria-popup subs-conf))]]
-                                   [ui/TableCell {:floated :left
-                                                  :width   2}
-                                    [:span
-                                     [ui/Checkbox {:key             "enable-new"
-                                                   :disabled        (empty? @notif-methods)
-                                                   :default-checked (:enabled subs-conf)
-                                                   :style           {:margin "1em"}
-                                                   :on-change       (ui-callback/checked
-                                                                      #(do
-                                                                         (dispatch-sync [::events/set-notification-subscription-config subs-conf])
-                                                                         (on-change :collection (:resource-kind subs-conf))
-                                                                         (on-change :enabled %)
-                                                                         (when (= 1 (count @notif-methods))
-                                                                           (on-change :method-id (-> @notif-methods
-                                                                                                     first
-                                                                                                     :id)))
-                                                                         (dispatch [::events/toggle-enabled (:id subs-conf) %])))}]]]
-                                   [ui/TableCell {:floated :left
-                                                  :width   4}
-                                    [SubsNotifMethodDropdown
-                                     (:method-ids subs-conf) notif-methods true (:resource-kind subs-conf) (:id subs-conf)]]
-
-                                   (let [subs-conf-id (:id subs-conf)]
-                                     [ui/TableCell {:floated :left
-                                                    :width   2
-                                                    :align   :right}
-                                      [uix/Button {:text     (@tr [:manage])
-                                                   :positive true
-                                                   :size     "small"
-                                                   :active   true
-                                                   :on-click #(do
-                                                                (dispatch [::events/set-notification-subscription-config-id subs-conf-id])
-                                                                (dispatch [::events/get-notification-subscriptions subs-conf-id])
-                                                                (dispatch [::events/open-notification-subscription-modal (get @subs-by-parent subs-conf-id)]))}]
-
-                                      [:span ff/nbsp ff/nbsp [ui/Label {:circular true} (get @subs-by-parent-counts subs-conf-id 0)]]])
-
-                                   [ui/TableCell {:floated :right
-                                                  :width   1
-                                                  :align   :right}
-                                    (when (general-utils/can-delete? subs-conf)
-                                      [DeleteButtonSubscriptionConfig subs-conf])
-                                    (when (general-utils/can-edit? subs-conf)
-                                      [ui/Icon {:name     :cog
-                                                :color    :blue
-                                                :style    {:cursor :pointer}
-                                                :on-click #(dispatch [::events/open-edit-subscription-config-modal subs-conf])}])]]))]]
+                         (for [subs-conf resource-subs-confs]
+                           ^{:key subs-conf}
+                           [ui/TableRow
+                            [ui/TableCell {:floated :left
+                                           :width   2}
+                             [:span (:name subs-conf)]]
+                            [ui/TableCell {:width 2} (criteria-popup subs-conf)]
+                            [ui/TableCell {:floated :left
+                                           :width   2}
+                             [:span
+                              [ui/Checkbox {:key             "enable-new"
+                                            :disabled        (empty? @notif-methods)
+                                            :default-checked (:enabled subs-conf)
+                                            :style           {:margin "1em"}
+                                            :on-change       (ui-callback/checked
+                                                              #(do
+                                                                 (dispatch-sync [::events/set-notification-subscription-config subs-conf])
+                                                                 (on-change :collection (:resource-kind subs-conf))
+                                                                 (on-change :enabled %)
+                                                                 (when (= 1 (count @notif-methods))
+                                                                   (on-change :method-id (-> @notif-methods
+                                                                                             first
+                                                                                             :id)))
+                                                                 (dispatch [::events/toggle-enabled (:id subs-conf) %])))}]]]
+                            [ui/TableCell {:floated :left
+                                           :width   4}
+                             [SubsNotifMethodDropdown
+                              (:method-ids subs-conf) notif-methods true (:resource-kind subs-conf) (:id subs-conf)]]
+                            [ui/TableCell {:floated :right
+                                           :width   1
+                                           :align   :right}
+                             (when (general-utils/can-delete? subs-conf)
+                               [DeleteButtonSubscriptionConfig subs-conf])
+                             (when (general-utils/can-edit? subs-conf)
+                               [ui/Icon {:name     :cog
+                                         :color    :blue
+                                         :style    {:cursor :pointer}
+                                         :on-click #(dispatch [::events/open-edit-subscription-config-modal subs-conf])}])]])]]
                        :title-size :h4
                        :default-open (= 0 idx)
                        :count (count resource-subs-confs)