+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Paginations - Items per page is stored to and loaded from localStorage #1009
 - Main - Remove deprecated message and bootstrap-message
 - Readme - Update dev section to include skip verification of untrusted certificate
+- Notifications - Added params to define time window reset and device name 
 
 ## [2.32.11] - 2022-12-19